chore: move logs to common
