+++ src/common/utils/log.ts
@@ -1,0 +1,6 @@
+export function log(text: string) {
+  if (!log.enabled) return
+  console.log(`[COW TRADING SDK] ${text}`)
+}
+
+log.enabled = false
